--- src/shared/Apis/login.ts
@@ -1,18 +1,0 @@
-import { devServerInstance } from "@shared/apiInstance";
-
-interface LoginRequestBody {
-  email: string;
-  password: string;
-}
-
-interface LoginResponse {
-  message: string;
-  token: string;
-}
-
-const postLogin = async (body: LoginRequestBody): Promise<LoginResponse> => {
-  const res = await devServerInstance.post("/api/login", { data: body });
-  return res.data;
-};
-
-export { postLogin}
